// generated/orval/fetch/google-health-api/models/exercise.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */
import type { ExerciseEvent } from './exerciseEvent';
import type { ExerciseExerciseType } from './exerciseExerciseType';
import type { ExerciseMetadata } from './exerciseMetadata';
import type { MetricsSummary } from './metricsSummary';
import type { SessionTimeInterval } from './sessionTimeInterval';
import type { SplitSummary } from './splitSummary';

/**
 * An exercise that stores information about a physical activity.
 */
export interface Exercise {
  /** Optional. Duration excluding pauses. */
  activeDuration?: string;
  /** Output only. Represents the timestamp of the creation of the exercise. */
  readonly createTime?: string;
  /** Required. The localized, human-readable name of the exercise. For all exercise types other than `OTHER`, the system ignores client input and overrides this field with a generated name based on `exercise_type` (e.g., "Walking" for `WALKING`). If `exercise_type` is `OTHER`, this field can contain the user's custom, free-form display name. */
  displayName?: string;
  /** Optional. Exercise events that happen during an exercise, such as pause & restarts. */
  exerciseEvents?: ExerciseEvent[];
  /** Optional. Additional exercise metadata. */
  exerciseMetadata?: ExerciseMetadata;
  /** Required. The type of activity performed during an exercise. */
  exerciseType?: ExerciseExerciseType;
  /** Required. Observed exercise interval */
  interval?: SessionTimeInterval;
  /** Required. Summary metrics for this exercise ( ) */
  metricsSummary?: MetricsSummary;
  /** Optional. Standard free-form notes captured at manual logging. */
  notes?: string;
  /** Optional. The default split is 1 km or 1 mile. - if the movement distance is less than the default, then there are no splits - if the movement distance is greater than or equal to the default, then we have splits */
  splits?: SplitSummary[];
  /** Optional. Laps or splits recorded within an exercise. Laps could be split based on distance or other criteria (duration, etc.) Laps should not be overlapping with each other. */
  splitSummaries?: SplitSummary[];
  /** Output only. This is the timestamp of the last update to the exercise. */
  readonly updateTime?: string;
}